Attempt to get smart_open to work with the S3 bucket
We were running into issues with smart_open not actually writing anything to the S3 bucket, and it appears to be due to the fact that the S3 Key object is not being handled properly.  However, we are currently unsure why that is the case because the downloads code has been working fine in this same capacity.  It is possible it may have something to do with the write mode; we had to change it from "wb" to just "w" for this to work, but it is still using "wb" under the hood when you step through the calls.

This commit does also update smart_open to the latest version, so we will want to make sure we test the download functionality as well to make sure it is still working properly after the change.