Add support for expression-based sorts in API
This changeset adds support for expression-based sorts in API queries, meaning that instead of sorting on a specific column we can choose to sort on an expression instead.  This is useful if an expression needs to be used to provide the desired value as opposed to what is stored in the column itself, especially if that expression is also used in the WHERE clause of the query.  Furthermore, this allows us to leverage expression-based indexes properly to ensure that the API queries are performant.

To create expression-based sorting options, follow the pattern developed for the sort_expressions hybrid property on the model in this PR.