Add remaining Schedule B COALESCE-based indexes
This changeset adds a set of COALESCE-based indexes in the Schedule B partition to augment the other ones we already have to cover the other permutations of queries that would have a COALECSE statement in them.

We are finding that the queries are still not as performant as they could be so we want to be sure these use cases are fully covered before moving on to other performance tuning opportunities.