Add more robust itemized schedule nightly processing
This changeset introduces more resilience to the nightly processing of the itemized schedule data by checking for a specific error condition and attempting to fix it before trying to process the data again.  We find that at times we run into a situation where an itemized record will be present in the new queue table as well as the master table, which indicates the need for an update, but the record is absent from the old queue table.  This change rectifies that by retrieving the values from the master table and creating a copy in the old queue, then retrying the processing again.  If the processing succeeds or fails in a fatal way (another type of error entirely or an error in trying to copy the record into the old queue), the loop is broken from the processing so the rest of the nightly refresh can continue.