Update Swagger UI to the latest release
This changeset finishes our migration away from the old custom forked version of Swagger UI that we had been using before and puts the newest version available in its place out of the box.  There are a few things that we need to adjust and we may want to incorporate some of the customizations we had before, so this will be a work in progress for a bit of time.

Furthermore, I was not sure how best to manage the static assets.  For now I have just included the whole static directory as is, but perhaps we just want to keep the directory itself and not its contents, and have Swagger UI built with every deploy.  This would mean that a developer would have to do this locally too as a part of project setup, though.  I wanted to gather input from the team first before settling on a solution to this.

The high-level changes include the following:

* Removed existing ./static/ and ./swagger-ui/ directories
* Removed gulpfile.js
* Removed all dependencies from package.json and replaced them with just swagger-ui-dist
* Rewrote all of the npm build scripts to account for the swagger-ui-dist package
* Adjusted the API error handling slightly to log the actual error happening
* Replaced most of the HTML in the ./webservices/templates/swagger-ui.html with the HTML from the new version of Swagger UI
* Updated flask-apispec to the lateset release